# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Invoice SaaS API",
        "description": "API REST pour la plateforme de facturation multi-tenant Invoice SaaS. Authentification via Laravel Sanctum (Bearer Token).",
        "contact": {
            "name": "Invoice SaaS Support",
            "email": "support@invoice-saas.com"
        },
        "license": {
            "name": "Propriétaire",
            "url": "https://invoice-saas.com/terms"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Serveur API principal"
        }
    ],
    "paths": {
...
        "/v1/invoices/{id}": {
            "get": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Afficher une facture",
                "operationId": "d1db505615d55efcff88ee0cd057e2be",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la facture",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la facture avec client et items",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "number": {
                                            "type": "string",
                                            "example": "INV-2025-0001"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "draft"
                                        },
                                        "subtotal": {
                                            "type": "number",
                                            "example": 500000
                                        },
                                        "tax": {
                                            "type": "number",
                                            "example": 90000
                                        },
                                        "total": {
                                            "type": "number",
                                            "example": 585000
                                        },
                                        "due_date": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "client": {
                                            "type": "object"
                                        },
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Facture non trouvée"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Mettre à jour une facture",
                "operationId": "3f5d8a391b679b4156bd1c7f9f9b1bf2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la facture",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "client_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-09-01"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "draft",
                                            "sent",
                                            "viewed",
                                            "partially_paid",
                                            "paid",
                                            "overdue"
                                        ],
                                        "example": "sent"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Mise à jour des conditions"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Facture mise à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invoice updated successfully"
                                        },
                                        "invoice": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Facture non trouvée"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Supprimer une facture",
                "operationId": "309e38a585791fb99c28a94f6a7f6e49",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la facture",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Facture supprimée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invoice deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Facture non trouvée"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/invoices/{id}/pdf": {
            "post": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Générer le PDF d'une facture",
                "operationId": "b513c7beb4f7d58539fe3498d4198b95",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la facture",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF généré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "pdf_path": {
                                                    "type": "string",
                                                    "example": "invoices/invoice-INV-2025-0001.pdf"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "PDF generated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Facture non trouvée"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}